Change the packaging structure to feature-based Create Dog entity Create DogPhoto entity Create ActivityLevel, Gender, PhotoType enums
