--- src/main/java/com/mateuszmarcyk/walk_the_dog/WalkTheDogApplication.java
@@ -1,13 +1,0 @@
-package com.mateuszmarcyk.walk_the_dog;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class WalkTheDogApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(WalkTheDogApplication.class, args);
-	}
-
-}
