--- src/main/java/com/mateuszmarcyk/walk_the_dog/controller/UserController.java
@@ -1,23 +1,0 @@
-package com.mateuszmarcyk.walk_the_dog.controller;
-
-import com.mateuszmarcyk.walk_the_dog.model.User;
-import com.mateuszmarcyk.walk_the_dog.service.UserService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
-
-@RequiredArgsConstructor
-@RestController
-@RequestMapping("/users")
-public class UserController {
-
-private final UserService userService;
-
-    @GetMapping
-    public List<User> findAll() {
-        return userService.getAll();
-    }
-}
