--- src/main/java/com/mateuszmarcyk/walk_the_dog/event/RegistrationCompleteEvent.java
@@ -1,20 +1,0 @@
-package com.mateuszmarcyk.walk_the_dog.event;
-
-import com.mateuszmarcyk.walk_the_dog.model.User;
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.context.ApplicationEvent;
-
-@Getter
-@Setter
-public class RegistrationCompleteEvent extends ApplicationEvent {
-
-    private User user;
-    private String applicationUrl;
-
-    public RegistrationCompleteEvent(User user, String applicationUrl) {
-        super(user);
-        this.user = user;
-        this.applicationUrl = applicationUrl;
-    }
-}
