--- src/main/java/com/mateuszmarcyk/walk_the_dog/exception/UserAlreadyExistsException.java
@@ -1,8 +1,0 @@
-package com.mateuszmarcyk.walk_the_dog.exception;
-
-public class UserAlreadyExistsException extends RuntimeException{
-
-    public UserAlreadyExistsException(String message) {
-        super(message);
-    }
-}
