--- src/main/java/com/mateuszmarcyk/walk_the_dog/model/User.java
@@ -1,54 +1,0 @@
-package com.mateuszmarcyk.walk_the_dog.model;
-
-
-import jakarta.persistence.*;
-import lombok.*;
-import org.hibernate.annotations.NaturalId;
-
-@Getter
-@Setter
-@ToString
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-@Table(name="app_user")
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id")
-    private Long id;
-
-    @Column(name = "first_name")
-    private String firstName;
-
-    @Column(name = "last_name")
-    private String lastName;
-
-
-    @Column(name = "username", unique = true)
-    private String username;
-
-    @NaturalId(mutable = true)
-    @Column(name = "email")
-    private String email;
-
-    @Column(name = "password")
-    private String password;
-
-    @Column(name = "phone_number")
-    private String phoneNumber;
-
-    @Column(name = "bio")
-    private String bio;
-
-    @Column(name = "profile_image_url")
-    private String profileImageUrl;
-
-    @Column(name = "app_user_role")
-    private String appUserRole;
-
-
-    @Column(name = "enabled")
-    private Boolean enabled = false;
-}
