--- src/main/java/com/mateuszmarcyk/walk_the_dog/model/UserRole.java
@@ -1,6 +1,0 @@
-package com.mateuszmarcyk.walk_the_dog.model;
-
-public enum UserRole {
-    ADMIN,
-    USER
-}
