--- src/main/java/com/mateuszmarcyk/walk_the_dog/registration/token/VerificationTokenRepository.java
@@ -1,8 +1,0 @@
-package com.mateuszmarcyk.walk_the_dog.registration.token;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface VerificationTokenRepository extends JpaRepository<VerificationToken, Long> {
-
-    VerificationToken findByToken(String token);
-}
