--- src/main/java/com/mateuszmarcyk/walk_the_dog/security/UserRegistrationDetails.java
@@ -1,65 +1,0 @@
-package com.mateuszmarcyk.walk_the_dog.security;
-
-import com.mateuszmarcyk.walk_the_dog.model.User;
-import lombok.Data;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.List;
-import java.util.stream.Collectors;
-
-@Data
-public class UserRegistrationDetails implements UserDetails {
-
-    private String username;
-    private String password;
-    private boolean isEnabled;
-    private List<GrantedAuthority> authorities;
-
-    public UserRegistrationDetails(User user) {
-        this.username = user.getEmail();
-        this.password = user.getPassword();
-        this.isEnabled = user.getEnabled();
-        this.authorities = Arrays.stream(user.getAppUserRole().split(","))
-                .map(SimpleGrantedAuthority::new)
-                .collect(Collectors.toList());
-    }
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return authorities;
-    }
-
-    @Override
-    public String getPassword() {
-        return password;
-    }
-
-    @Override
-    public String getUsername() {
-        return username;
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return isEnabled;
-    }
-}
