--- src/main/java/com/mateuszmarcyk/walk_the_dog/service/UserService.java
@@ -1,20 +1,0 @@
-package com.mateuszmarcyk.walk_the_dog.service;
-
-import com.mateuszmarcyk.walk_the_dog.model.User;
-import com.mateuszmarcyk.walk_the_dog.registration.RegistrationRequest;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface UserService {
-
-    List<User> getAll();
-
-    User register(RegistrationRequest request);
-
-    Optional<User> findByEmail(String email);
-
-    void saveVerificationToken(User user, String verificationToken);
-
-    String validateToken(String token);
-}
