+++ src/main/java/com/mateuszmarcyk/walkthedog/WalkTheDogApplication.java
@@ -1,0 +1,13 @@
+package com.mateuszmarcyk.walkthedog;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class WalkTheDogApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(WalkTheDogApplication.class, args);
+	}
+
+}
