+++ src/main/java/com/mateuszmarcyk/walkthedog/dog/ActivityLevel.java
@@ -1,0 +1,7 @@
+package com.mateuszmarcyk.walkthedog.dog;
+
+public enum ActivityLevel {
+    LOW,
+    MEDIUM,
+    HIGH
+}
