+++ src/main/java/com/mateuszmarcyk/walkthedog/dog/Gender.java
@@ -1,0 +1,6 @@
+package com.mateuszmarcyk.walkthedog.dog;
+
+public enum Gender {
+    MALE,
+    FEMALE
+}
