+++ src/main/java/com/mateuszmarcyk/walkthedog/dogphoto/PhotoType.java
@@ -1,0 +1,6 @@
+package com.mateuszmarcyk.walkthedog.dogphoto;
+
+public enum PhotoType {
+    REGULAR,
+    PROFILE
+}
