+++ src/main/java/com/mateuszmarcyk/walkthedog/event/RegistrationCompleteEvent.java
@@ -1,0 +1,20 @@
+package com.mateuszmarcyk.walkthedog.event;
+
+import com.mateuszmarcyk.walkthedog.user.User;
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.context.ApplicationEvent;
+
+@Getter
+@Setter
+public class RegistrationCompleteEvent extends ApplicationEvent {
+
+    private User user;
+    private String applicationUrl;
+
+    public RegistrationCompleteEvent(User user, String applicationUrl) {
+        super(user);
+        this.user = user;
+        this.applicationUrl = applicationUrl;
+    }
+}
