+++ src/main/java/com/mateuszmarcyk/walkthedog/exception/UserAlreadyExistsException.java
@@ -1,0 +1,8 @@
+package com.mateuszmarcyk.walkthedog.exception;
+
+public class UserAlreadyExistsException extends RuntimeException{
+
+    public UserAlreadyExistsException(String message) {
+        super(message);
+    }
+}
