+++ src/main/java/com/mateuszmarcyk/walkthedog/registration/RegistrationRequest.java
@@ -1,0 +1,11 @@
+package com.mateuszmarcyk.walkthedog.registration;
+
+public record RegistrationRequest(
+        String firstName,
+        String lastName,
+        String username,
+        String email,
+        String password,
+        String role
+) {
+}
