+++ src/main/java/com/mateuszmarcyk/walkthedog/registration/token/VerificationTokenRepository.java
@@ -1,0 +1,8 @@
+package com.mateuszmarcyk.walkthedog.registration.token;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface VerificationTokenRepository extends JpaRepository<VerificationToken, Long> {
+
+    VerificationToken findByToken(String token);
+}
