+++ src/main/java/com/mateuszmarcyk/walkthedog/security/UserRegistrationSecurityConfig.java
@@ -1,0 +1,47 @@
+package com.mateuszmarcyk.walkthedog.security;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.web.SecurityFilterChain;
+
+import static org.springframework.security.config.Customizer.withDefaults;
+
+@Configuration
+@EnableWebSecurity
+public class UserRegistrationSecurityConfig {
+
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+        return new BCryptPasswordEncoder();
+    }
+
+    @Bean
+    public SecurityFilterChain securityFilterChain(HttpSecurity httpSecurity) throws Exception {
+        Object AbstractHttpConfigurer;
+        httpSecurity
+                .cors(withDefaults())
+                .csrf(configurer -> configurer.disable())
+                .authorizeHttpRequests(authorizeRequests -> authorizeRequests
+                        .requestMatchers("/register/**").permitAll()
+                        .requestMatchers("/users/**").hasAnyAuthority("USER", "ADMIN")
+                ).formLogin();
+//                .formLogin(formLogin -> formLogin
+//                        .loginPage("/login")
+//                        .loginProcessingUrl("/process")
+//                        .defaultSuccessUrl("/users")
+//                        .permitAll()
+//                ).logout(logout ->
+//                        logout.permitAll()
+//                );
+
+        return httpSecurity.build();
+    }
+
+
+}
+
+
