+++ src/main/java/com/mateuszmarcyk/walkthedog/user/UserController.java
@@ -1,0 +1,21 @@
+package com.mateuszmarcyk.walkthedog.user;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
+
+@RequiredArgsConstructor
+@RestController
+@RequestMapping("/users")
+public class UserController {
+
+private final UserService userService;
+
+    @GetMapping
+    public List<User> findAll() {
+        return userService.getAll();
+    }
+}
