+++ src/main/java/com/mateuszmarcyk/walkthedog/user/UserRepository.java
@@ -1,0 +1,10 @@
+package com.mateuszmarcyk.walkthedog.user;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    Optional<User> findByEmail(String email);
+}
