+++ src/main/java/com/mateuszmarcyk/walkthedog/user/UserRole.java
@@ -1,0 +1,6 @@
+package com.mateuszmarcyk.walkthedog.user;
+
+public enum UserRole {
+    ADMIN,
+    USER
+}
