--- src/test/java/com/mateuszmarcyk/walk_the_dog/WalkTheDogApplicationTests.java
@@ -1,13 +1,0 @@
-package com.mateuszmarcyk.walk_the_dog;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class WalkTheDogApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
