+++ src/test/java/com/mateuszmarcyk/walkthedog/WalkTheDogApplicationTests.java
@@ -1,0 +1,13 @@
+package com.mateuszmarcyk.walkthedog;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class WalkTheDogApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
